requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
audioop-lts==0.2.2
blinker==1.9.0
buildozer==1.5.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distlib==0.4.0
docutils==0.22
filelock==3.18.0
filetype==1.2.0
Flask==3.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
Kivy==2.3.1
Kivy-examples==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.gstreamer==0.3.4
kivy_deps.sdl2==0.8.0
MarkupSafe==3.0.2
packaging==25.0
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pefile==2023.2.7
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
pywin32==311
pywin32-ctypes==0.2.3
requests==2.32.4
rsa==4.9.1
screeninfo==0.8.1
setuptools==80.9.0
sh==2.2.2
SpeechRecognition==3.14.3
sswg==4.3.0
standard-aifc==3.13.0
standard-chunk==3.13.0
sudo==1.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
ursina==8.1.1
virtualenv==20.33.1
Werkzeug==3.1.3
